connection-streamlit/requirements.txt
snowflake-connector-python
snowflake-snowpark-python==1.29.0
streamlit